Changes to allow batch parsing and writing of larger files

+++ Ebcdic2Unicode/ParserUtilities.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections;
+using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using System.Text.RegularExpressions;
@@ -215,9 +216,9 @@ namespace Ebcdic2Unicode
             return true;
         }
 
-        public static bool WriteParsedLineArrayToTxt(ParsedLine[] lines, string outputFilePath, string delimiter = "\t", bool includeColumnNames = true, bool addQuotes = true)
+        public static bool WriteParsedLineArrayToTxt(ParsedLine[] lines, string outputFilePath, string delimiter = "\t", bool includeColumnNames = true, bool addQuotes = true, string quoteCharacter = "\"", bool append = false)
         {
-            Console.WriteLine("{0}: Writing output TXT file...", DateTime.Now);
+            Console.WriteLine("{0}: Writing output to TXT file...", DateTime.Now);
 
             if (lines == null || lines.Length == 0)
             {
@@ -227,20 +228,20 @@ namespace Ebcdic2Unicode
 
             try
             {
-                if (File.Exists(outputFilePath))
+                if (File.Exists(outputFilePath) && !append)
                 {
                     File.Delete(outputFilePath);
                 }
 
                 using (TextWriter tw = new StreamWriter(outputFilePath, true))
                 {
-                    if (includeColumnNames && lines.Length > 0)
+                    if (includeColumnNames && lines.Length > 0 & !append)
                     {
-                        tw.WriteLine(string.Join(delimiter, lines[0].Template.GetFieldNames(addQuotes)));
+                        tw.WriteLine(string.Join(delimiter, lines[0].Template.GetFieldNames(addQuotes, quoteCharacter)));
                     }
                     foreach (ParsedLine line in lines)
                     {
-                        tw.WriteLine(string.Join(delimiter, line.GetFieldValues(addQuotes)));
+                        tw.WriteLine(string.Join(delimiter, line.GetFieldValues(addQuotes, quoteCharacter)));
                     }
                 }
             }
@@ -250,13 +251,13 @@ namespace Ebcdic2Unicode
                 return false;
             }
 
-            Console.WriteLine("{1}: Output file created {0}", Path.GetFileName(outputFilePath), DateTime.Now);
+            Console.WriteLine($"{DateTime.Now}: Output file {(append ? "written to" : "created")} {Path.GetFileName(outputFilePath)}");
             return true;
         }
 
-        public static bool WriteParsedLineArrayToCsv(ParsedLine[] lines, string outputFilePath, bool includeColumnNames = true, bool addQuotes = true)
+        public static bool WriteParsedLineArrayToCsv(ParsedLine[] lines, string outputFilePath, bool includeColumnNames = true, bool addQuotes = true, bool append = false)
         {
-            return WriteParsedLineArrayToTxt(lines, outputFilePath, ",", includeColumnNames, addQuotes);
+            return WriteParsedLineArrayToTxt(lines, outputFilePath, ",", includeColumnNames, addQuotes, "\"" ,append);
         }
 
         public static string RemoveNonAsciiChars(string text)
@@ -301,6 +302,27 @@ namespace Ebcdic2Unicode
 
             result = text.Substring(text.Length - length);
             return result;
+        }
+
+        public static List<byte[]> Slice(this byte[] source, int length, int chunkSize)
+        {
+            int recordCount = (source.Length / length);
+            int chunkCount = 0;
+            do
+            {
+                chunkCount++;
+            }
+            while (chunkCount * chunkSize < recordCount);
+
+            List<byte[]> slices = new List<byte[]>();
+            for (int i = 1; i <= chunkCount; i++)
+            {
+                byte[] slice = new byte[length * chunkSize];
+                Array.Copy(source, i * length, slice, 0, length * chunkSize);
+                slices.Add(slice);
+            }
+            
+            return slices;
         }
     }
 }
